requirements.txt
# Root requirements file for Azure deployment
# Azure looks for this in the root directory

# Backend dependencies
jedi==0.18.1
tornado==6.1
packaging
ptyprocess==0.7.0
terminado==0.18.1
matplotlib>=3.10.0
numpy>=2.3.0
bcrypt==4.3.0
requests==2.32.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
psutil==5.9.8